Move `ManualMap` into `Matches` lint pass
